requirements.txt
amqp==5.3.1
asgiref==3.9.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.14.2
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.1
curl_cffi==0.13.0
daphne==4.2.1
Django==5.2.6
djangorestframework==3.16.1
frozendict==2.4.6
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
msgpack==1.1.1
multitasking==0.0.12
numpy==2.3.3
packaging==25.0
pandas==2.3.3
peewee==3.18.2
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
yfinance==0.2.66
zope.interface==8.0.1
